remote hook fixes; int3 remote breakpoint fixed

+++ src/BlackBone/Subsystem/x86Subsystem.cpp
@@ -132,7 +132,7 @@ ptr_t x86Native::getTEB( HANDLE hThread, _TEB32* pteb )
     _THREAD_BASIC_INFORMATION_T<DWORD> tbi = { 0 };
     ULONG bytes = 0;
 
-    if (NT_SUCCESS( SAFE_NATIVE_CALL( NtQueryInformationThread, hThread, (THREADINFOCLASS)0, &tbi, sizeof( tbi ), &bytes ) ) && pteb)
+    if (NT_SUCCESS( SAFE_NATIVE_CALL( NtQueryInformationThread, hThread, (THREADINFOCLASS)0, &tbi, (ULONG)sizeof( tbi ), &bytes ) ) && pteb)
         ReadProcessMemory( _hProcess, reinterpret_cast<LPCVOID>(tbi.TebBaseAddress), pteb, sizeof(_TEB32), NULL );
 
     return tbi.TebBaseAddress;
